requirements.txt
python-telegram-bot
google-generativeai
google-cloud-speech
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
beautifulsoup4
newspaper3k
feedparser
schedule